+++ aldrin/CHANGELOG.md
@@ -23,6 +23,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add `low_level::PendingReply::id()`.
 - Add `low_level::Call::take_args()`.
 - Add `low_level::Event::take_args()`.
+- Add `low_level::Reply::take_args()`.
 
 ### Changed